modify(feat/sang/#89): 주소 검색창 팝업 함수 라이브러리로 구현
